requirements.txt
google-generativeai>=0.7.2
python-dotenv>=1.0.1
numpy>=1.24.0
scikit-learn>=1.3.0
